also extract IntegrationTestKitBase
